Support Zoom, add some logic fixes
1) Add Zoom to the supported clients.

2) There are two kinds of clients we need to handle. One, for example
Zoom, disconnects from both the microphone and the speakers when it
isn't active. The other, for example Google Chrome, might be connected
to the speakers for output when it isn't using the microphone. We need
to handle the two kinds of clients differently in terms of deciding
when to switch from HSP back to A2DP.

3) We need to handle better the kinds of clients that use a different
name for connecting to the speakers vs. the microphone.