Handle Zoom better, wait for quiescence before checking state changes
Zoom is annoyingly complicated. Sometimes it hangs onto the speaker
connection when it isn't in a call, so we have to treat it as a
persistent client. Worse, sometimes it hangs onto both the speaker
*and* microphone when it isn't in a call, in which case there really
isn't much we can do except adding a comment to the script alerting
people to this problem and what to do about it.

Rather than checking whether we should switch profiles immediately
when we get an event from `pactl`, wait until things have been client
for a second before doing any checks, to give clients time to
establish or remove both of their connections before we mess with
them. This prevents flapping.

Populate the list of existing clients on startup so the script does
the right thing if a client is already connected to the speaker or
microphone before it starts up.